Refactor toolbar settings and adjust GUI element focus
In this commit, the toolbar settings were moved from being global to being nested within the MainWnd struct. Additionally, the condition that checks for the .Name flag in bool.swg was updated. This condition also now adjusts the extension of focus for GUI elements when the flag is not present. GUI dimensions are also updated based on the toolbar labels' visibility.