Update Scope functionality and add Public Implementation handling
This commit includes multiple changes to enhance our scope handling functionality. Notably, it introduces the ability to handle public implementations, with the addition of a PublicImpl Set into the PublicSet struct. Furthermore, it contains changes in the Scope functions addPublicFunc and addPublicImpl to account for this new feature. In the process of upgrading this functionality, there have been changes in various files, including adjustments to FormatAst, Ast, and Syntax files.