Refactor fillpath and painter with tessellate instance.
The code in fillpath.swg and painter.swg was simplified by using a persistent Poly2Tri.Tessellate instance (named `tess`), stored in the painter, instead of creating a new instance for every polylines computation. This changes aims to optimise the performance by reusing the tessellate instance and reducing the complexity of the 'fillpath' function.