Update variable g_Page and refactor tessellation structures
The value of variable g_Page in the draw module has been changed from 1 to 7. The structures for tessellation in the poly2tri subsystem of the pixel module have been reorganized for better code clarity and maintainability. Certain structural declarations were moved to a more accurate position in the code and some adjustments were also made to the addPolyLine method.