Add optimization function to reduce stack instructions
The changes include the addition of a new function, reduceStack1(), to optimize stack instructions further in ByteCodeOptimizer. This new function optimizes the replication of 'GetFromStack' and 'SetAtPointer' operations by consolidating into 'GetFromStack16/32/64' and 'SetAtPointer16/32/64' operations based on certain conditions, thereby simplifying the bytecode execution and improving the performance. It also includes a minor change commenting out a portion of code for future reference.