Add replaceAll method to Utf8 class and update ByteCode print
A `replaceAll` method has been added to `Utf8` class that replaces all occurrences of a substring with another one. The signature of `printSourceCode` function in `ByteCode` has been updated to include an extra `lastInline` parameter of type `AstNode**`. This update is reflected in the function's definition and its calls in `ByteCode_Print.cpp`.