Update stack size in emitLocalFuncDecl
The code that ensures the updated stack size because of late inline operators has been shifted to the beginning of the emitLocalFuncDecl function. This commit is intended to update the DecSPBP value right after verifying changes in the stack size rather than after emitting the debug line and byte operation.