Add visibility control for split pane items
This update introduces a new feature which allows us to control the visibility of individual panes in a split view. The "visible" property has been added to each split view item, linked to showPane and isPaneVisible functions. An example usage of these functions has been added in the wnd module.