Add visibility management to splitter control
The splitter control in the GUI module was improved to include pane visibility management. A new field, `numVisiblePanes`, was added to keep track of the visible panes. Changes also were made to methods such as `addPane` and `showPane` to accommodate this feature.  In addition, modifications were made to the noise example to adjust the presentation.