Refactor property attributes and improve child event handling
The property attributes have been moved to a new file 'propertyattributes.swg' from 'propertylist.swg'. This change simplifies the 'propertylist.swg' file and improves the structure of the codebase. In addition, handling of the child events in the GUI has been improved by adding a 'hookOnChildEvent' hook for processing any event of a child window. This allows for more precise event handling and control of child elements in the GUI.