Add tooltip feature to property list items
The changes included adding tooltip functionality to property list items. The tooltips display the name of the item when the mouse hovers over it. A condition to hide the tooltip when the mouse leaves the item was also incorporated. A commented out line that hides the label description was also added.