Add seed parameter to Perlin and Value noise methods
The seed parameter was added to the Perlin and Value noise methods in the FastNoise module. This change will allow the noise generation to be more customizable and potentially more diverse. This adds an additional layer of randomness to the noise generation process by giving control over the seed of the random number generation.