Refactor theme color properties and update tab push button dimensions
Refactored the naming convention of several theme color properties in `themecolors.swg` and `properties.swg` to be more consistent. Modified the behavior of background coloring in the grid type. Additionally, in `themerects.swg`, the dimensions for push button components under the 'tabItem' class have been updated to improve visual alignment.