Add focus control for specific elements and fix window unregistering
The commit extends the ability to control focus changes for sections and categories in `properties.swg` file. Also, it refines the method of unregistering windows in `application.swg`. Now, besides checking if the window is the same, it also checks if the window is the owner of the mouse or keyboard focus window.