Fix bug to auto detect an enum value if inside an attribute use
The check in Semantic_Identifier is broadened to include 'Attribute' symbols, not only 'Function' and 'Variable'. This effectively enhances symbol handling for better flexibility. Additionally, the use of `Swag.Warning` in various files is adjusted by using the more concise format `.Enable` and `.Disable` instead of the long `Swag.WarningLevel.Enable` and `Swag.WarningLevel.Disable`.