Add functionality to print bytecode source when debugging. Debug "set" command
This commit introduces a new feature in the bytecode debugger wherein the source can be printed along with bytecodes. Moreover, the option printBcCode has been added to the set commands, allowing the user to toggle this feature on and off. The code blocks that print instructions have been adjusted to respect this new setting.