Revised command syntax and updated execution for ByteCode Debugger
The syntax of various commands in the ByteCode Debugger has been updated, using "--" instead of "/" for command options. Changes have been made to the execution of these commands to align with the syntax update. Additionally, amended the output message in the "set" command to clarify that it prints also the source code when printing bytecode.