Refactor GUI property module for margin adjustments
Removed the `leftMargin` from column size calculations and made other relevant adjustments in the GUI properties module. This includes changes in the resizing of grid columns and the properties of items like the setting of `resizingColumn` to false on `MouseReleased`. Also ensured the margins are set correctly for various items and conditions.