Update variable declaration in array.swg
Changed the variable 'arrData' in file '../bin/std/modules/gui/src/property/values/array.swg' from 'var' to 'let' as it is not being re-assigned after initial setting. This adds better readability to the code and follows best practice for variable declaration when a variable does not need to change its value.