Replace HasChilds flag with numChilds property
This update replaces the HasChilds flag with the numChilds variable in order to count the number of children for each item. This simplification has led to changes in several conditions that originally checked for the presence of children. The addition of children is now managed through a separate function, thereby avoiding code duplication.