Update GUI elements destruction and array reconstruction algorithms
This commit includes the necessary updates in GUI elements destruction and array reconstruction methods. In the GUI window module, the child elements are ensured to be destroyed along with the parent element. In the array module, array items and child items are properly restructured and removed, while keeping track of state updates. There are also improvements in casting the count of array data to user data.