Add level parameter to appendTypedValue methods
The appendTypedValue and appendTypedValueProtected methods in ByteCodeDebugger interface have been updated to include an additional parameter, level. This new parameter has been used in multiple function calls, controlling the visibility of structure and array printing depending on the recursion level. This update refines the codebase to be more customizable in terms of the debugging options it provides.