Refine error handling in ByteCodeDebugger
This commit improves error handling and feedback in ByteCodeDebugger. It changes the return value from 'BadArguments' to 'NotEnoughArguments' or 'Error'. It also adds checks for evaluation of dynamic expressions and prints an error message when expression to evaluate is empty, or the print format cannot be applied to the type.