Update modal loop tracking in menu control and property updates
This commit adds tracking of the modal loop state to prevent inappropriate property updating while in the loop. Changes have been made to 'menuctrl.swg' to increment and decrement the 'inModalLoop' counter, and in 'properties.swg' to make the updating of properties contingent on the 'inModalLoop' state. Also, some minor changes have been made to fix parameter mismatch and potential null typeStruct.