Refactor error handling and update SwagContext size
This commit revises the handling of errors in the Module_Run module by moving the error check to the end of the execution. A 'pushHasError' attribute has been added to the SwagContext, and its size has been updated to 1600 in LLVM.cpp. Moreover, minor formatting changes were applied to ByteCode_Print for better readability.