Update icon set and simplify array construction
The code updates the icon set for the "Remove" option and makes changes to the actions for deleting captures and selections. These changes include setting the icon for these actions. On a different note, it also significantly simplifies the construction of arrays in the values/array file by reducing the code and refreshing the array in a more straightforward way.