Add selection handling to properties GUI and fix array check
A new field 'selItem' has been added to keep track of the currently selected item in the properties GUI. Selection handling has been implemented during mouse events and theme color application is now based on this selection. In addition, a null-check has been added before checking if a type is a structure in the 'isStructArray' function within array.swg.