Add day9B function to solve Advent of Code problem
This commit introduces the function day9B which solves part two of day 9 problem for Advent of Code 2022. The function includes a simulation for a longer rope with ten knots. It is added to the main execution code for the overall set of problems. The solution was confirmed with a successful assertion for the expected count.