Refactor code for bytecode optimization
The import of "MapRegTo.h" has been reshuffled in "ByteCodeOptContext.h" for better code structure. A new variable "vecInstCopy" is introduced to improve handling of bytecode instructions in the optimization process. Several instances of "vecInst1" have been replaced with "vecInstCopy" to reflect this change, improving consistency and readability of the bytecode optimizing code.