Refactor and enhance ByteCodeDebugger functionality
The ByteCodeDebugger code was modified for better management and display of breakpoints. A new function 'printBreakpoint' was added which has improved the visualization of breakpoints, especially for line breakpoints in source code. Moreover, 'printSourceLines' function was updated to prevent potential null pointer dereference. Invalid range lines now returns an error.