Implement one-shot breakpoint feature
The changes introduce the functionality of one-shot breakpoints in the ByteCode debugger. A one-shot breakpoint, represented by "tb" or "tbreak" commands, is a type of breakpoint that would automatically get removed after being hit. Additionally, updates were made to ensure the count of breakpoints does not surpass the limit of 8.