Update breakpoint messages in ByteCode debugger
This commit changes the wording in messages related to breakpoints in the ByteCode debugger. Specifically, it alters the description when a memory breakpoint change is detected and also refines the display format for file and line breakpoints. These changes aim to provide clearer and more concise information to the user.