Simplify argument handling in ByteCodeDebugger
The ByteCodeDebugger's handling of arguments in list and longlist commands was simplified. The change made the function only accepts zero or one argument, eliminating the need for 'findCmdBc' function. This leads to cleaner, more readable code and fewer potential points of failure. The function [name] params in 'longlist' and 'longinst' of the command list were also removed.