Add location parameter to ByteCodeDebugger print functions
The commit introduces a new location parameter to the printTitleNameType function in the ByteCodeDebugger class. This function and several calling functions have been renamed to printTitleNameTypeLoc to reflect the added functionality. It enables printing the location details along with the title, name, and type, providing more contextual information in the console logs.