Add solution for Advent of Code 2022 day 10
This commit includes the solution for day 10 of Advent of Code 2022. The program reads from a file called 10.txt and follows the instructions within to calculate required problems. It applies different strategies based on the current cycle of a virtual CPU. The solution is divided into part A and B, each with its unique logic implementation.