Refactor byte code handling and debugging
This update changes several aspects of byte code handling and debugging. It modifies parameter name for clarity and adds 'AstNode' member to 'Location' struct to improve source code location tracking. Also, it simplifies the 'cmdLongList' method in 'ByteCodeDebugger', reducing the complexity of source code retrieval. This refactoring results in cleaner code with improved readability and maintainability.