Refactor method call and debug mode setting in ByteCodeDebugger
The method call 'safeOwnerInline()' has been replaced with 'hasOwnerInline()' in ByteCodeDebugger.cpp, to better reflect the function's purpose. The setting and resetting of the 'bcMode' debug mode in ByteCodeDebugger_List.cpp has been rearranged for more logical code flow. Before, the mode was set to false at the start of some functions and now it is set to false at the end.