Add force option to doSyntaxColor function and update calls
The `doSyntaxColor` function has been updated to include a new `force` option. This parameter allows syntax coloring to be performed even if the global options are set to disable it. This is particularly useful for situations such as generating documentation where the default coloring options should not apply. The calls to `doSyntaxColor` throughout the codebase have been updated to use this new option where appropriate. Also, the use of "logColors" has been simplified to "logColor".