Optimize ClearMaskU32 operation in ByteCodeOptimizer
This commit adds optimization for the ClearMaskU32 operation in the ByteCodeOptimizer_Reduce.cpp file. Particularly, it ensures that if the ClearMaskU32 and ClearMaskU64 operations have the same 'a' value and the ClearMaskU64 operation doesn't start a statement, the 'b' value is updated to be the minimum of the two 'b' values. Also, a minor formatting change was made to align the equals sign in some assignments.