Refactor FFI function addressing logic
The previous implementation of obtaining an FFI function's address has been removed from the ByteCodeRun class and instead incorporated it into the AstNode. This will simplify the codebase by reducing the duplication of this logic, and make the function addressing process more efficient by storing the address within the AstNode itself. Adjustments were also made to the bytecode operations related to foreign calls.