Add LocalCallPop0Param2 and ForeignCallPop0Param2 bytecode operations
These operations support calling local and foreign functions with no stack adjustments after the call, while having two parameters pushed onto the stack. The operations are optimized in the ByteCodeOptimizer_ReduceX2 function and implemented in the run function. The support for these operations is also added to the LLVM and SCBE backends.