Refactor syntax and bytecode operation codes
This commit primarily refactors syntax and bytecode operations in both LanguageSpec.cpp and ByteCodeOpList.h. Changes made include simplifying code by transforming conditions from "jump if" style to "if ... jump" format, as well as adding "call" as a new keyword. This is part of an effort to make our codebase clearer and more maintainable.