Add [[nodiscard]] attribute to multiple methods
The [[nodiscard]] attribute has been added to multiple methods across various files, signaling that the return value of these methods is not to be ignored. This improves code reliability by suggesting that the compiler issues a warning if the returned value is not used. The rename from `getGlobalPrivFct` to `getGlobalPrivateFct` in several files for improved readability is also included.