Add [[nodiscard]] attribute to Workspace methods
The [[nodiscard]] attribute has been added to several methods in the Workspace structure to encourage checking of return values. This is expected to help to prevent errors resulting from overlooked return values, thus improving the robustness and reliability of the code. Also, the order of some methods is changed for better readability and structure.