Add [[nodiscard]] attribute to relevant functions
The [[nodiscard]] attribute has been added to functions where it makes sense to use the returned value, which can avoid bugs associated with overlooking a returned value. The correction ensures developers are aware that they should be utilizing the returned value for these functions. Furthermore, a cast from unused-result of markDebugBlock to void was added to suppress warnings.