Add new Space Invaders game script
This commit covers the addition of a new Space Invaders game script. It updates and corrects the path variables for loading and binding textures for the game. It also amends test script batch files to include the Space Invaders script to be tested. Minor changes to other scripts like the Pacman game script were also made to correct loading and binding textures.